+++ Modules/xNetworking/DSCResources/MSFT_xDNSServerAddress/MSFT_xDNSServerAddress.psm1
@@ -1,36 +1,27 @@
-#######################################################################################
-#  xDNSServerAddress : DSC Resource that will set/test/get the current DNS Server
-#  Address, by accepting values among those given in xDNSServerAddress.schema.mof
-#######################################################################################
+# Get the path to the shared modules folder
+$script:ModulesFolderPath = Join-Path -Path (Split-Path -Path (Split-Path -Path $PSScriptRoot -Parent)) `
+                                      -ChildPath 'Modules'
 
-data LocalizedData
-{
-    # culture="en-US"
-    ConvertFrom-StringData -StringData @'
-GettingDNSServerAddressesMessage=Getting the DNS Server Addresses.
-ApplyingDNSServerAddressesMessage=Applying the DNS Server Addresses.
-DNSServersSetCorrectlyMessage=DNS Servers are set correctly.
-DNSServersAlreadySetMessage=DNS Servers are already set correctly.
-CheckingDNSServerAddressesMessage=Checking the DNS Server Addresses.
-DNSServersNotCorrectMessage=DNS Servers are not correct. Expected "{0}", actual "{1}".
-DNSServersHaveBeenSetCorrectlyMessage=DNS Servers were set to the desired state.
-InterfaceNotAvailableError=Interface "{0}" is not available. Please select a valid interface and try again.
-AddressFormatError=Address "{0}" is not in the correct format. Please correct the Address parameter in the configuration and try again.
-AddressIPv4MismatchError=Address "{0}" is in IPv4 format, which does not match server address family {1}. Please correct either of them in the configuration and try again.
-AddressIPv6MismatchError=Address "{0}" is in IPv6 format, which does not match server address family {1}. Please correct either of them in the configuration and try again.
-'@
-}
+# Import the Networking Resource Helper Module
+Import-Module -Name (Join-Path -Path $script:ModulesFolderPath `
+                               -ChildPath (Join-Path -Path 'NetworkingDsc.ResourceHelper' `
+                                                     -ChildPath 'NetworkingDsc.ResourceHelper.psm1'))
 
-######################################################################################
-# The Get-TargetResource cmdlet.
-# This function will get the present list of DNS ServerAddress DSC Resource
-# schema variables on the system
-######################################################################################
+# Import Localization Strings
+$script:localizedData = Get-LocalizedData `
+    -ResourceName 'MSFT_xDNSServerAddress' `
+    -ResourcePath $PSScriptRoot
+
+# Import the common networking functions
+Import-Module -Name (Join-Path -Path $script:ModulesFolderPath `
+                               -ChildPath (Join-Path -Path 'NetworkingDsc.Common' `
+                                                     -ChildPath 'NetworkingDsc.Common.psm1'))
+
 function Get-TargetResource
 {
     [OutputType([System.Collections.Hashtable])]
     param
-    (        
+    (
         [Parameter(Mandatory)]
         [ValidateNotNullOrEmpty()]
         [String[]]$Address,
@@ -43,7 +34,7 @@ function Get-TargetResource
         [ValidateSet('IPv4', 'IPv6')]
         [String]$AddressFamily
     )
-    
+
     Write-Verbose -Message ( @( "$($MyInvocation.MyCommand): "
         $($LocalizedData.GettingDNSServerAddressesMessage)
         ) -join '')
@@ -59,15 +50,11 @@ function Get-TargetResource
     $returnValue
 }
 
-######################################################################################
-# The Set-TargetResource cmdlet.
-# This function will set a new Server Address in the current node
-######################################################################################
 function Set-TargetResource
 {
     param
-    (    
-        #IP Address that has to be set    
+    (
+        #IP Address that has to be set
         [Parameter(Mandatory)]
         [ValidateNotNullOrEmpty()]
         [String[]]$Address,
@@ -79,7 +66,7 @@ function Set-TargetResource
         [Parameter(Mandatory)]
         [ValidateSet('IPv4', 'IPv6')]
         [String]$AddressFamily,
-        
+
         [Boolean]$Validate = $false
     )
 
@@ -114,8 +101,8 @@ function Set-TargetResource
             $($LocalizedData.DNSServersHaveBeenSetCorrectlyMessage)
             ) -join '' )
     }
-    else 
-    { 
+    else
+    {
         #Test will return true in this case
         Write-Verbose -Message ( @( "$($MyInvocation.MyCommand): "
             $($LocalizedData.DNSServersAlreadySetMessage)
@@ -123,15 +110,11 @@ function Set-TargetResource
     }
 }
 
-######################################################################################
-# The Test-TargetResource cmdlet.
-# This will test if the given Server Address is among the current node's Server Address collection
-######################################################################################
 function Test-TargetResource
 {
     [OutputType([System.Boolean])]
     param
-    (        
+    (
         [Parameter(Mandatory)]
         [ValidateNotNullOrEmpty()]
         [String[]]$Address,
@@ -143,8 +126,8 @@ function Test-TargetResource
         [Parameter(Mandatory)]
         [ValidateSet('IPv4', 'IPv6')]
         [String]$AddressFamily,
-        
-        [Boolean]$Validate = $false        
+
+        [Boolean]$Validate = $false
     )
     # Flag to signal whether settings are correct
     [Boolean] $desiredConfigurationMatch = $true
@@ -154,7 +137,7 @@ function Test-TargetResource
         ) -join '' )
 
     #Validate the Settings passed
-    Foreach ($ServerAddress in $Address) {       
+    Foreach ($ServerAddress in $Address) {
         Test-ResourceProperty `
             -Address $ServerAddress `
             -AddressFamily $AddressFamily `
@@ -181,8 +164,8 @@ function Test-TargetResource
                 -f ($Address -join ','),($currentAddress -join ',')
             ) -join '' )
     }
-    else 
-    { 
+    else
+    {
         #Test will return true in this case
         Write-Verbose -Message ( @( "$($MyInvocation.MyCommand): "
             $($LocalizedData.DNSServersSetCorrectlyMessage)
@@ -191,9 +174,6 @@ function Test-TargetResource
     return $desiredConfigurationMatch
 }
 
-#######################################################################################
-#  Helper functions
-#######################################################################################
 function Test-ResourceProperty
 {
     # Function will check the Address details are valid and do not conflict with
@@ -267,7 +247,5 @@ function Test-ResourceProperty
         $PSCmdlet.ThrowTerminatingError($errorRecord)
     }
 } # Test-ResourceProperty
-#######################################################################################
 
-#  FUNCTIONS TO BE EXPORTED 
 Export-ModuleMember -function Get-TargetResource, Set-TargetResource, Test-TargetResource