+++ Modules/xNetworking/DSCResources/MSFT_xDefaultGatewayAddress/MSFT_xDefaultGatewayAddress.psm1
@@ -1,38 +1,27 @@
-#######################################################################################
-#  xDefaultGatewayAddress : DSC Resource that will set/test/get the current default gateway
-#  Address, by accepting values among those given in xDefaultGatewayAddress.schema.mof
-#######################################################################################
+# Get the path to the shared modules folder
+$script:ModulesFolderPath = Join-Path -Path (Split-Path -Path (Split-Path -Path $PSScriptRoot -Parent)) `
+                                      -ChildPath 'Modules'
 
-data LocalizedData
-{
-    # culture="en-US"
-    ConvertFrom-StringData -StringData @'
-GettingDefaultGatewayAddressMessage=Getting the Default Gateway Address.
-ApplyingDefaultGatewayAddressMessage=Applying the Default Gateway Address.
-DefaultGatewayAddressSetToDesiredStateMessage=Default Gateway address was set to the desired state.
-DefaultGatewayRemovedMessage=Default Gateway address has been removed.
-CheckingDefaultGatewayAddressMessage=Checking the Default Gateway Address.
-DefaultGatewayNotMatchMessage=Default gateway does NOT match desired state. Expected "{0}", actual "{1}".
-DefaultGatewayCorrectMessage=Default gateway is correct.
-DefaultGatewayDoesNotExistMessage=Default gateway does not exist. Expected "{0}".
-DefaultGatewayExistsButShouldNotMessage=Default gateway exists but it should not.
-DefaultGatewayExistsAndShouldMessage=Default Gateway does not exist which is correct.
-InterfaceNotAvailableError=Interface "{0}" is not available. Please select a valid interface and try again.
-AddressFormatError=Address "{0}" is not in the correct format. Please correct the Address parameter in the configuration and try again.
-AddressIPv4MismatchError=Address "{0}" is in IPv4 format, which does not match server address family {1}. Please correct either of them in the configuration and try again.
-AddressIPv6MismatchError=Address "{0}" is in IPv6 format, which does not match server address family {1}. Please correct either of them in the configuration and try again.
-'@
-}
+# Import the Networking Resource Helper Module
+Import-Module -Name (Join-Path -Path $script:ModulesFolderPath `
+                               -ChildPath (Join-Path -Path 'NetworkingDsc.ResourceHelper' `
+                                                     -ChildPath 'NetworkingDsc.ResourceHelper.psm1'))
 
-######################################################################################
-# The Get-TargetResource cmdlet.
-# This function will get the current Default Gateway Address
-######################################################################################
+# Import Localization Strings
+$script:localizedData = Get-LocalizedData `
+    -ResourceName 'MSFT_xDefaultGateway' `
+    -ResourcePath $PSScriptRoot
+
+# Import the common networking functions
+Import-Module -Name (Join-Path -Path $script:ModulesFolderPath `
+                               -ChildPath (Join-Path -Path 'NetworkingDsc.Common' `
+                                                     -ChildPath 'NetworkingDsc.Common.psm1'))
+
 function Get-TargetResource
 {
     [OutputType([System.Collections.Hashtable])]
     param
-    (        
+    (
         [String]$Address,
 
         [Parameter(Mandatory)]
@@ -43,11 +32,11 @@ function Get-TargetResource
         [ValidateSet('IPv4', 'IPv6')]
         [String]$AddressFamily
     )
-    
+
     Write-Verbose -Message ( @("$($MyInvocation.MyCommand): "
         $($LocalizedData.GettingDefaultGatewayAddressMessage)
         ) -join '' )
-    
+
     # Use $AddressFamily to select the IPv4 or IPv6 destination prefix
     $DestinationPrefix = '0.0.0.0/0'
     if ($AddressFamily -eq 'IPv6')
@@ -74,11 +63,6 @@ function Get-TargetResource
     $returnValue
 }
 
-######################################################################################
-# The Set-TargetResource cmdlet.
-# This function will set the Default Gateway Address for the Interface/Family in the
-# current node
-######################################################################################
 function Set-TargetResource
 {
     param
@@ -94,7 +78,7 @@ function Set-TargetResource
         [String]$AddressFamily
     )
     # Validate the parameters
-    
+
     Write-Verbose -Message ( @("$($MyInvocation.MyCommand): "
         $($LocalizedData.ApplyingDefaultGatewayAddressMessage)
         ) -join '' )
@@ -147,11 +131,6 @@ function Set-TargetResource
     }
 }
 
-######################################################################################
-# The Test-TargetResource cmdlet.
-# This will test if the given Address is set as the Gateway Server address for the
-# Interface/Family in the current node
-######################################################################################
 function Test-TargetResource
 {
     [OutputType([System.Boolean])]
@@ -236,9 +215,6 @@ function Test-TargetResource
     return $desiredConfigurationMatch
 }
 
-#######################################################################################
-#  Helper functions
-#######################################################################################
 function Test-ResourceProperty {
     # Function will check the Address details are valid and do not conflict with
     # Address family. Ensures interface exists.
@@ -313,7 +289,5 @@ function Test-ResourceProperty {
         }
     }
 } # Test-ResourceProperty
-#######################################################################################
 
-#  FUNCTIONS TO BE EXPORTED 
 Export-ModuleMember -function Get-TargetResource, Set-TargetResource, Test-TargetResource