+++ Modules/xNetworking/DSCResources/MSFT_xHostsFile/MSFT_xHostsFile.psm1
@@ -1,19 +1,21 @@
-#region localizeddata
-if (Test-Path "${PSScriptRoot}\${PSUICulture}")
-{
-    Import-LocalizedData -BindingVariable LocalizedData `
-                         -Filename MSFT_xHostsFile.psd1 `
-                         -BaseDirectory "${PSScriptRoot}\${PSUICulture}"
-} 
-else
-{
-    #fallback to en-US
-    Import-LocalizedData -BindingVariable LocalizedData `
-                         -Filename MSFT_xHostsFile.psd1 `
-                         -BaseDirectory "${PSScriptRoot}\en-US"
-}
-#endregion
+# Get the path to the shared modules folder
+$script:ModulesFolderPath = Join-Path -Path (Split-Path -Path (Split-Path -Path $PSScriptRoot -Parent)) `
+                                      -ChildPath 'Modules'
 
+# Import the Networking Resource Helper Module
+Import-Module -Name (Join-Path -Path $script:ModulesFolderPath `
+                               -ChildPath (Join-Path -Path 'NetworkingDsc.ResourceHelper' `
+                                                     -ChildPath 'NetworkingDsc.ResourceHelper.psm1'))
+
+# Import Localization Strings
+$script:localizedData = Get-LocalizedData `
+    -ResourceName 'MSFT_xHostsFile' `
+    -ResourcePath $PSScriptRoot
+
+# Import the common networking functions
+Import-Module -Name (Join-Path -Path $script:ModulesFolderPath `
+                               -ChildPath (Join-Path -Path 'NetworkingDsc.Common' `
+                                                     -ChildPath 'NetworkingDsc.Common.psm1'))
 
 function Get-TargetResource
 {
@@ -22,11 +24,11 @@ function Get-TargetResource
         [Parameter(Mandatory = $true)]
         [System.String]
         $HostName,
-        
+
         [Parameter(Mandatory = $false)]
         [System.String]
         $IPAddress,
-        
+
         [Parameter(Mandatory = $false)]
         [System.String]
         [ValidateSet("Present","Absent")]
@@ -38,33 +40,33 @@ function Get-TargetResource
     $hosts = Get-Content -Path "$env:windir\System32\drivers\etc\hosts"
     $allHosts = $hosts `
            | Where-Object { [System.String]::IsNullOrEmpty($_) -eq $false -and $_.StartsWith('#') -eq $false } `
-           | ForEach-Object { 
+           | ForEach-Object {
                 $data = $_ -split '\s+'
-                if ($data.Length -gt 2) 
+                if ($data.Length -gt 2)
                 {
                     # Account for host entries that have multiple entries on a single line
                     $result = @()
-                    for ($i = 1; $i -lt $data.Length; $i++) 
+                    for ($i = 1; $i -lt $data.Length; $i++)
                     {
                         $result += @{
                             Host = $data[$i]
                             IP = $data[0]
-                        }    
+                        }
                     }
                     return $result
                 }
-                else 
+                else
                 {
                     return @{
                         Host = $data[1]
                         IP = $data[0]
-                    }    
+                    }
                 }
         } | Select-Object @{ Name="Host"; Expression={$_.Host}}, @{Name="IP"; Expression={$_.IP}}
-        
+
     $hostEntry = $allHosts | Where-Object { $_.Host -eq $HostName }
-    
-    if ($null -eq $hostEntry) 
+
+    if ($null -eq $hostEntry)
     {
         return @{
             HostName = $HostName
@@ -72,7 +74,7 @@ function Get-TargetResource
             Ensure = "Absent"
         }
     }
-    else 
+    else
     {
         return @{
             HostName = $hostEntry.Host
@@ -88,22 +90,22 @@ function Set-TargetResource
         [Parameter(Mandatory = $true)]
         [System.String]
         $HostName,
-        
+
         [Parameter(Mandatory = $false)]
         [System.String]
         $IPAddress,
-        
+
         [Parameter(Mandatory = $false)]
         [System.String]
         [ValidateSet("Present","Absent")]
         $Ensure = "Present"
     )
-    
+
     $currentValues = Get-TargetResource @PSBoundParameters
-    
+
     Write-Verbose -Message ($LocalizedData.StartingSet -f $HostName)
 
-    if ($Ensure -eq "Present" -and $PSBoundParameters.ContainsKey("IPAddress") -eq $false) 
+    if ($Ensure -eq "Present" -and $PSBoundParameters.ContainsKey("IPAddress") -eq $false)
     {
         $errorId = 'IPAddressNotPresentError'
         $errorCategory = [System.Management.Automation.ErrorCategory]::InvalidArgument
@@ -115,22 +117,22 @@ function Set-TargetResource
 
         $PSCmdlet.ThrowTerminatingError($errorRecord)
     }
-    
+
     if ($currentValues.Ensure -eq "Absent" -and $Ensure -eq "Present")
     {
         Write-Verbose -Message ($LocalizedData.CreateNewEntry -f $HostName)
         Add-Content -Path "$env:windir\System32\drivers\etc\hosts" -Value "`r`n$IPAddress`t$HostName"
     }
-    else 
+    else
     {
         $hosts = Get-Content -Path "$env:windir\System32\drivers\etc\hosts"
-        $replace = $hosts | Where-Object { 
-            [System.String]::IsNullOrEmpty($_) -eq $false -and $_.StartsWith('#') -eq $false 
+        $replace = $hosts | Where-Object {
+            [System.String]::IsNullOrEmpty($_) -eq $false -and $_.StartsWith('#') -eq $false
         } | Where-Object { $_ -like "*$HostName" }
 
         $multiLineEntry = $false
         $data = $replace -split '\s+'
-        if ($data.Length -gt 2) 
+        if ($data.Length -gt 2)
         {
             $multiLineEntry = $true
         }
@@ -138,26 +140,26 @@ function Set-TargetResource
         if ($currentValues.Ensure -eq "Present" -and $Ensure -eq "Present")
         {
             Write-Verbose -Message ($LocalizedData.UpdateExistingEntry -f $HostName)
-            if ($multiLineEntry -eq $true) 
+            if ($multiLineEntry -eq $true)
             {
                 $newReplaceLine = $replace -replace $HostName, ""
                 $hosts = $hosts -replace $replace, $newReplaceLine
                 $hosts += "$IPAddress`t$HostName"
             }
-            else 
+            else
             {
-                $hosts = $hosts -replace $replace, "$IPAddress`t$HostName"    
+                $hosts = $hosts -replace $replace, "$IPAddress`t$HostName"
             }
         }
         if ($Ensure -eq "Absent")
         {
             Write-Verbose -Message ($LocalizedData.RemoveEntry -f $HostName)
-            if ($multiLineEntry -eq $true) 
+            if ($multiLineEntry -eq $true)
             {
                 $newReplaceLine = $replace -replace $HostName, ""
                 $hosts = $hosts -replace $replace, $newReplaceLine
             }
-            else 
+            else
             {
                 $hosts = $hosts -replace $replace, ""
             }
@@ -173,26 +175,26 @@ function Test-TargetResource
         [Parameter(Mandatory = $true)]
         [System.String]
         $HostName,
-        
+
         [Parameter(Mandatory = $false)]
         [System.String]
         $IPAddress,
-        
+
         [Parameter(Mandatory = $false)]
         [System.String]
         [ValidateSet("Present","Absent")]
         $Ensure = "Present"
     )
-    
+
     $currentValues = Get-TargetResource @PSBoundParameters
     Write-Verbose -Message ($LocalizedData.StartingTest -f $HostName)
-    
-    if ($Ensure -ne $currentValues.Ensure) 
+
+    if ($Ensure -ne $currentValues.Ensure)
     {
         return $false
     }
-   
-    if ($Ensure -eq "Present" -and $IPAddress -ne $currentValues.IPAddress) 
+
+    if ($Ensure -eq "Present" -and $IPAddress -ne $currentValues.IPAddress)
     {
         return $false
     }