+++ Modules/xNetworking/DSCResources/MSFT_xIPAddress/MSFT_xIPAddress.psm1
@@ -1,35 +1,22 @@
-#######################################################################################
-#  xIPAddress : DSC Resource that will set/test/get the current IP
-#  Address, by accepting values among those given in xIPAddress.schema.mof
-#######################################################################################
+# Get the path to the shared modules folder
+$script:ModulesFolderPath = Join-Path -Path (Split-Path -Path (Split-Path -Path $PSScriptRoot -Parent)) `
+                                      -ChildPath 'Modules'
 
-data LocalizedData
-{
-    # culture="en-US"
-    ConvertFrom-StringData -StringData @'
-GettingIPAddressMessage=Getting the IP Address.
-ApplyingIPAddressMessage=Applying the IP Address.
-IPAddressSetStateMessage=IP Interface was set to the desired state.
-CheckingIPAddressMessage=Checking the IP Address.
-IPAddressDoesNotMatchMessage=IP Address does NOT match desired state. Expected {0}, actual {1}.
-IPAddressMatchMessage=IP Address is in desired state.
-PrefixLengthDoesNotMatchMessage=Prefix Length does NOT match desired state. Expected {0}, actual {1}.
-PrefixLengthMatchMessage=Prefix Length is in desired state.
-DHCPIsNotDisabledMessage=DHCP is NOT disabled.
-DHCPIsAlreadyDisabledMessage=DHCP is already disabled.
-DHCPIsNotTestedMessage=DHCP status is ignored when Address Family is IPv6.
-InterfaceNotAvailableError=Interface "{0}" is not available. Please select a valid interface and try again.
-AddressFormatError=Address "{0}" is not in the correct format. Please correct the Address parameter in the configuration and try again.
-AddressIPv4MismatchError=Address "{0}" is in IPv4 format, which does not match server address family {1}. Please correct either of them in the configuration and try again.
-AddressIPv6MismatchError=Address "{0}" is in IPv6 format, which does not match server address family {1}. Please correct either of them in the configuration and try again.
-PrefixLengthError=A Prefix Length of {0} is not valid for {1} addresses. Please correct the Prefix Length and try again.
-'@
-}
+# Import the Networking Resource Helper Module
+Import-Module -Name (Join-Path -Path $script:ModulesFolderPath `
+                               -ChildPath (Join-Path -Path 'NetworkingDsc.ResourceHelper' `
+                                                     -ChildPath 'NetworkingDsc.ResourceHelper.psm1'))
 
-######################################################################################
-# The Get-TargetResource cmdlet.
-# This function will get the present list of IP Address DSC Resource schema variables on the system
-######################################################################################
+# Import Localization Strings
+$script:localizedData = Get-LocalizedData `
+    -ResourceName 'MSFT_xIPAddress' `
+    -ResourcePath $PSScriptRoot
+
+# Import the common networking functions
+Import-Module -Name (Join-Path -Path $script:ModulesFolderPath `
+                               -ChildPath (Join-Path -Path 'NetworkingDsc.Common' `
+                                                     -ChildPath 'NetworkingDsc.Common.psm1'))
+
 function Get-TargetResource
 {
     [OutputType([System.Collections.Hashtable])]
@@ -67,10 +54,6 @@ function Get-TargetResource
     $returnValue
 }
 
-######################################################################################
-# The Set-TargetResource cmdlet.
-# This function will set a new IP Address in the current node
-######################################################################################
 function Set-TargetResource
 {
     param
@@ -158,10 +141,6 @@ function Set-TargetResource
         ) -join '' )
 } # Set-TargetResource
 
-######################################################################################
-# The Test-TargetResource cmdlet.
-# This will test if the given IP Address is among the current node's IP Address collection
-######################################################################################
 function Test-TargetResource
 {
     [OutputType([System.Boolean])]
@@ -248,9 +227,6 @@ function Test-TargetResource
     return $desiredConfigurationMatch
 } # Test-TargetResource
 
-#######################################################################################
-#  Helper functions
-#######################################################################################
 function Test-ResourceProperty {
     # Function will check the IP Address details are valid and do not conflict with
     # Address family. Also checks the prefix length and ensures the interface exists.
@@ -346,7 +322,5 @@ function Test-ResourceProperty {
     }
 
 } # Test-ResourceProperty
-#######################################################################################
 
-#  FUNCTIONS TO BE EXPORTED
 Export-ModuleMember -function Get-TargetResource, Set-TargetResource, Test-TargetResource