+++ Modules/xNetworking/DSCResources/MSFT_xRoute/MSFT_xRoute.psm1
@@ -1,27 +1,21 @@
-data LocalizedData
-{
-    # culture="en-US"
-    ConvertFrom-StringData -StringData @'
-GettingRouteMessage=Getting {0} Route on "{1}" dest {2} nexthop {3}.
-RouteExistsMessage={0} Route on "{1}" dest {2} nexthop {3} exists.
-RouteDoesNotExistMessage={0} Route on "{1}" dest {2} nexthop {3} does not exist.
-SettingRouteMessage=Setting {0} Route on "{1}" dest {2} nexthop {3}.
-EnsureRouteExistsMessage=Ensuring {0} Route on "{1}" dest {2} nexthop {3} exists.
-EnsureRouteDoesNotExistMessage=Ensuring {0} Route on "{1}" dest {2} nexthop {3} does not exist.
-RouteCreatedMessage={0} Route on "{1}" dest {2} nexthop {3} has been created.
-RouteUpdatedMessage={0} Route on "{1}" dest {2} nexthop {3} has been updated.
-RouteRemovedMessage={0} Route on "{1}" dest {2} nexthop {3} has been removed.
-TestingRouteMessage=Testing {0} Route on "{1}" dest {2} nexthop {3}.
-RoutePropertyNeedsUpdateMessage={4} property on {0} Route on "{1}" dest {2} nexthop {3} is different. Change required.
-RouteDoesNotExistButShouldMessage={0} Route on "{1}" dest {2} nexthop {3} does not exist but should. Change required.
-RouteExistsButShouldNotMessage={0} Route on "{1}" dest {2} nexthop {3} exists but should not. Change required.
-RouteDoesNotExistAndShouldNotMessage={0} Route on "{1}" dest {2} nexthop {3} does not exist and should not. Change not required.
-InterfaceNotAvailableError=Interface "{0}" is not available. Please select a valid interface and try again.
-AddressFormatError=Address "{0}" is not in the correct format. Please correct the Address parameter in the configuration and try again.
-AddressIPv4MismatchError=Address "{0}" is in IPv4 format, which does not match address family {1}. Please correct either of them in the configuration and try again.
-AddressIPv6MismatchError=Address "{0}" is in IPv6 format, which does not match address family {1}. Please correct either of them in the configuration and try again.
-'@
-}
+# Get the path to the shared modules folder
+$script:ModulesFolderPath = Join-Path -Path (Split-Path -Path (Split-Path -Path $PSScriptRoot -Parent)) `
+                                      -ChildPath 'Modules'
+
+# Import the Networking Resource Helper Module
+Import-Module -Name (Join-Path -Path $script:ModulesFolderPath `
+                               -ChildPath (Join-Path -Path 'NetworkingDsc.ResourceHelper' `
+                                                     -ChildPath 'NetworkingDsc.ResourceHelper.psm1'))
+
+# Import Localization Strings
+$script:localizedData = Get-LocalizedData `
+    -ResourceName 'MSFT_xRoute' `
+    -ResourcePath $PSScriptRoot
+
+# Import the common networking functions
+Import-Module -Name (Join-Path -Path $script:ModulesFolderPath `
+                               -ChildPath (Join-Path -Path 'NetworkingDsc.Common' `
+                                                     -ChildPath 'NetworkingDsc.Common.psm1'))
 
 function Get-TargetResource
 {